+++ src/tui/layout.rs
@@ -19,6 +19,8 @@ use super::app::Page;
 use super::pages::browse::BrowsePage;
 use super::pages::export::ExportPage;
 use super::pages::import::ImportPage;
+use super::pages::login::LoginPage;
+use super::pages::logout::LogoutPage;
 use super::pages::pull::PullPage;
 use super::pages::settings::SettingsPage;
 use super::pages::status::StatusPage;
@@ -31,6 +33,8 @@ pub struct PageSet {
     pub browse: BrowsePage,
     pub export: ExportPage,
     pub watch: WatchPage,
+    pub login: LoginPage,
+    pub logout: LogoutPage,
     pub settings: SettingsPage,
 }
 
@@ -43,6 +47,8 @@ impl PageSet {
             browse: BrowsePage::default(),
             export: ExportPage::default(),
             watch: WatchPage::default(),
+            login: LoginPage::default(),
+            logout: LogoutPage::default(),
             settings: SettingsPage::default(),
         }
     }
@@ -55,6 +61,8 @@ impl PageSet {
             Page::Browse => &self.browse,
             Page::Export => &self.export,
             Page::Watch => &self.watch,
+            Page::Login => &self.login,
+            Page::Logout => &self.logout,
             Page::Settings => &self.settings,
         }
     }
@@ -67,6 +75,8 @@ impl PageSet {
             Page::Browse => &mut self.browse,
             Page::Export => &mut self.export,
             Page::Watch => &mut self.watch,
+            Page::Login => &mut self.login,
+            Page::Logout => &mut self.logout,
             Page::Settings => &mut self.settings,
         }
     }